Update kit2_quickstart.md
The phrase "- CM-550 controller is initially programmed with Kit2 example (`MAX-E2`, `Commando`, `Scorpi`, `AutoBot`,`Strider`)." doesn't make sense. If the CM-550 comes with ENG KIT 1, how can it come pre-programmed with KIT 2 examples? I'm guessing this is trying to let users know that CM-550 comes with preinstalled examples for KIT 1, but the examples for KIT 2 need to be installed.